Fix aircrack-ng fdsan error.
So... long story short:
https://github.com/pitube08642/aircrack-ng-for-termux/issues/13

Cracking .cap handshake handy with crunch or similar programm.
So usually we do something like this:
$ crunch 8 8 0123 | aircrack-ng -b <MAC> dump-01.cap
piping crunch output to the aircrack.
But in termux aircrack-ng v1.7 raises error:
fdsan: failed to exchange ownership of file descriptor:
fd 0 is owned by FILE* 0x7625720028, was expected to be unowned

This fix is similar to termux cgdb (thanks Henrik)
https://github.com/termux/termux-packages/commit/3abcd55aae5f9b180a3bff7e87671088bad59aa1

---------------------------------------------
Date:      Tue Feb 18 23:44:37 2025 +0300

On branch bugfix/termux_fdsan_error
Changes to be committed:
	modified:   src/aircrack-ng/aircrack-ng.c